Fix pg_regress parameter '--psqldir' in 6X_STABLE
The pg_regress parameter '--psqldir' intends to switch psqldir to let users
change the bindir of the postgres. But not all 'bindir's are replaced with
'psqldir', so if the pg_regress is configured and maked in dir1, and the
modules users try to run pg_regress with --psqldir=dir2, if the dir1 cannot
be found for some reasons(maybe renamed in pipeline), it will print exits
with "sh: /usr/local/greenplum-db-devel/bin/psql: No such file or directory"
error messages.

So this pr solution is to replace psqldir with bindir, so that if '--psqldir'
is set, then all the 'bindir' is set to a new dir so that all command line
tools not only psql are set to a new bindir.
(partially cherry-pick from dcae5faccab64776376d354decda0017c648bb53)