Fix gpfdist writable transform table cannot report error.
1. Modify function write_subprocess to report error.
2. Fix header_callback to update response content.
3. Save errfile for writable tranformations.

For writable external tables, gpfdist protocol uses X-GP-PROTO:0,
so it is hard to report error for segments. Yet we can use INTERNAL_ERROR
to report(function check_response). The file->http_response is updated
by header_callback, but the content can only be updated only once. So
we change the code to update the file->http_response if the header line
starts with "HTTP/".

Issue: #9497