Fix pgbouncer regression when login as admin users.
The previous commit d83d349 may cause a regression that
when admin users login into pgbouncer database, the login
user will be switched to pgbouncer which is not a good way
 for "hba" or other auth_type to login into pgbouncer database.

If you want to login into pgbouncer database as admin_user when
auth_type is set to "hba", you need to add an entry for the
admin user in auth_hba_file and auth_file. Also if you set the
login rule as md5 or other authentication ways which need
password, you need to set the password in auth_file.