Fix custom formatter handling logic.
The function externalgettup_custom() is to handle custom formatter
logic when loading data with external tables. But if custom protocol
like PXF is used to loading data, the EOF is not reached when the
last content of data is returned, the EOF is reached only when the
last reading returns 0. So in the current logic of function
externalgettup_custom(), it may lose the chance for custom formatter
to handle the last line of data.

This commit removes the using of raw_buf_len, instead the real data
provides for custom formatter to use is formatter->fmt_databuf. So
we check the length this buffer to give custom formatter last chance
to handle the last line of data. Even if the last line is not
complete, it is the custom formatter's responsibility to report this.