Add detect infinite loop logic to avoid inifite loop.
The right way to return FMT_NEED_MORE_DATA in custom formatter is to
check EOF flag before return. But currently, a lot of custom formatter
does not implement in the right way. They just return FMT_NEED_MORE_DATA
when they need more data in the current block of line without changing
the formatter->fmt_databuf. So it will cause infinite loop.

In order to prevent infinite loop, we detect this by adding a count when
eof flag is set. So this PR enables custom protocol to handle the last
line of incomplete data if the custom formatter has taken care of the eof
flag, and also prevent infinite loop regression.