Gpfdist allow client side retry when failed because of network failure.
There are retry mechanism on the gpdb segment side which is as client
to gpfdist server when writing into writable external tables. However,
the retry is not working because of gpfdist legacy code that will
close fstream when failed to receive data from socket.

This commit allow the client side retry sending post request.