Add session clean up timeout parameter for gpfdist with '-k' option.
The default value for session clean up timeout is 300s, but in heavy
workload, this default value is not long enough to wait for segment
next POST request to arrive. Then it will cause "400 invalid request
due to wrong sequence number" failure.

So this commit add one option '-k' to set this value for gpfdist.
The default is 300, the valid range would be (300..86400).

Conflicts:
	src/bin/gpfdist/gpfdist.c