Add size range for write_file_size to be more safe.
The write_file_buffer is used by all POST request. The start buffer size will be 1M.
The maximum size will be 1G. We want all POST data fit in the write_file_buffer, so
that when network error happens, no data will be written into the fstream.
The sending buffer at segment size will be range from 32K to 128M, so 1G is large
enough to hold all the POST data.

If the Content-length is larger then 1G, an 400 will be returned.