Enable two-phase commit for foreign table.
For postgres_fdw, it supports 2PC. But the 2PC was closed for histroy
reason, so we need to open it for postgres_fdw.
When foreign table connects to different remote servers, it would be
possible that some remote update succeed, but some does not. We need
to ensure the atomicity of this scenario.

Here is how it is implemented:

The legacy implementation of postgres_fdw has registered one callback
function called `pgfdw_xact_callback` to handle this transaction issue.
After we enable two-phase commit for foreign table, this could be
ensured as we tested.

Yet the commit is implemented by SQL sentences, so if the execution of
"commit" failed on some QEs, it could still cause inconsistency. And
this issue could not be solved by now.