This reverts commit f1e36291f0add4c4aace321886c73c5baf05a141.
Here explains why we need this modfication for file syncrep.c:
In each walsender process, there is one structure pointer MyWalSnd which
points to the structure of WalSnd, and if the walsender is a phsical
walsender, then the MyWalSnd->is_for_gp_walreceiver is set to true.
In function SyncRepWaitForLSN(), if the process is forked from coordinator
postgres, then it tries to wait for candidate whose walsnd->is_for_gp_walreceiver
is set to true which means it only wants to wait for the lsn to be confirmed by
the physical walreceiver, not including logical walreceiver.
But in coordinator, in logical walsender process, the value of
MyWalSnd->sync_standby_priority is 0, which means not listed. And in function
SyncRepGetSyncRecPtr()->SyncRepGetCandidateStandbys(), it will search all of
the standby candidates, and if logical walsender process is selected, that
means the updates of value of WalSndCtl->lsn[1] will be effected by logical
walsender, and for coordinator, the update of lsn is slow, so it will cause
the updating of the shared memory value WalSndCtl->lsn[1] slowly, then in
function SyncRepWaitForLSN(), it will wait for the slowest updating lsn to be
updated which is updated in logical walsender which is not what we expect.
So we need to add one another condition for searching the candicate standby
in function SyncRepGetCandidateStandbys().